@@ -14,8 +14,8 @@
         }
     ],
     "require": {
-        "php": ">=5.3.3",
-        "zf-commons/zfc-user": "~1.2"
+        "php": ">=7.1",
+        "lm-commons/lmc-user": "3.5.0"
     },
     "autoload": {
         "psr-0": {
