@@ -10,26 +10,7 @@ use ZfcUserImpersonate\Controller;
 
 return array(
     'factories' => array(
-        'zfcuserimpersonate_adminController' => function ($cm) {
-            $sm = $cm->getServiceLocator();
-
-            $adminController = new Controller\Admin();
-            $adminController->setConfig($sm->get('zfcuserimpersonate_module_options'));
-            $adminController->setUserService($sm->get('zfcuserimpersonate_user_service'));
-
-            return $adminController;
-        },
-        'zfcuser' => function($cm) {
-            /* @var ControllerManager $cm*/
-            $serviceManager = $cm->getServiceLocator();
-
-            /* @var RedirectCallback $redirectCallback */
-            $redirectCallback = $serviceManager->get('zfcuser_redirect_callback');
-
-            /* @var UserController $controller */
-            $controller = new Controller\User($redirectCallback);
-
-            return $controller;
-        },
+        'zfcuserimpersonate_adminController' => Controller\Factory\AdminFactory::class,
+        'zfcuser' => Controller\Factory\UserFactory::class,
     )
 );
