@@ -6,20 +6,9 @@
  * @author Mark Tudor <code AT icefusion DOT co DOT uk>
  */
 
-use ZfcUserImpersonate\View\Helper\ZfcUserImpersonatorDisplayName;
-use ZfcUserImpersonate\View\Helper\ZfcUserImpersonatorIdentity;
-
 return array(
     'factories' => array(
-        'zfcUserImpersonatorDisplayName' => function ($hm) {
-            $viewHelper = new ZfcUserImpersonatorDisplayName();
-            $viewHelper->setUserService($hm->getServiceLocator()->get('zfcuserimpersonate_user_service'));
-            return $viewHelper;
-        },
-        'zfcUserImpersonatorIdentity' => function ($hm) {
-            $viewHelper = new ZfcUserImpersonatorIdentity();
-            $viewHelper->setUserService($hm->getServiceLocator()->get('zfcuserimpersonate_user_service'));
-            return $viewHelper;
-        },
+        'zfcUserImpersonatorDisplayName' => \ZfcUserImpersonate\View\Helper\Factory\ZfcUserImpersonatorDisplayNameFactory::class,
+        'zfcUserImpersonatorIdentity' => \ZfcUserImpersonate\View\Helper\Factory\ZfcUserImpersonatorIdentityFactory::class,
     ),
 );
