@@ -10,9 +10,9 @@
 
 namespace ZfcUserImpersonate\Service;
 
-use Zend\Authentication\Storage\StorageInterface;
-use ZfcUser\Entity\UserInterface;
-use ZfcUser\Service\User as ZfcUserUserService;
+use Laminas\Authentication\Storage\StorageInterface;
+use LmcUser\Entity\UserInterface;
+use LmcUser\Service\User as ZfcUserUserService;
 use ZfcUserImpersonate\Exception\Domain as DomainException;
 use ZfcUserImpersonate\Exception\NotImpersonating as NotImpersonatingException;
 use ZfcUserImpersonate\Exception\UserNotFound as UserNotFoundException;
@@ -24,7 +24,7 @@ class User extends ZfcUserUserService
      * The storage container in which the 'impersonator' (real user) is stored whilst they are impersonating another
      * user.
      *
-     * @var \Zend\Authentication\Storage\StorageInterface
+     * @var \Laminas\Authentication\Storage\StorageInterface
      */
     protected $storageForImpersonator;
 
@@ -60,7 +60,8 @@ class User extends ZfcUserUserService
         }
 
         // Store the 'impersonator' (real user) in storage to allow later unimpersonation.
-        $this->getStorageForImpersonator()->write($this->getAuthService()->getIdentity());
+        $id = $this->getAuthService()->getIdentity()->getId();
+        $this->getStorageForImpersonator()->write($id);
 
         // Config setting determines whether to write the whole object to the session
         // or just the ID
@@ -87,26 +88,10 @@ class User extends ZfcUserUserService
         }
 
         // Retrieve the 'impersonator' (real user) from storage.
-        $impersonatorUser = $this->getStorageForImpersonator()->read();
-
-        // Assert that the 'impersonator' (real user) is valid.
-        if (!$impersonatorUser instanceof UserInterface) {
-            // The 'impersonator' (real user) is not the correct type.
-            throw new DomainException(
-                '$$impersonatorUser is not an instance of UserInterface',
-                500
-            );
-        }
-
-        // Config setting determines whether to write the whole object to the session
-        // or just the ID
-        if (!$this->getStoreUserAsObject()) {
-            $impersonatorUser = $impersonatorUser->getId();
-        }
-
+        $impersonatorUserId = $this->getStorageForImpersonator()->read();
 
         // End impersonation by restoring the original identity - the 'impersonator' (real user) - to auth storage.
-        $this->getAuthService()->getStorage()->write($impersonatorUser);
+        $this->getAuthService()->getStorage()->write($impersonatorUserId);
 
         // Clear the 'impersonator' (real user) from storage.
         $this->getStorageForImpersonator()->clear();
@@ -128,7 +113,7 @@ class User extends ZfcUserUserService
      *
      * Session storage is used by default unless a different storage adapter has been set.
      *
-     * @return \Zend\Authentication\Storage\StorageInterface
+     * @return StorageInterface
      */
     public function getStorageForImpersonator()
     {
@@ -140,8 +125,8 @@ class User extends ZfcUserUserService
      *
      * Session storage is used by default unless a different storage adapter has been set.
      *
-     * @param  \Zend\Authentication\Storage\StorageInterface $storageForImpersonator
-     * @return \ZfcUser\Service\User
+     * @param  StorageInterface $storageForImpersonator
+     * @return User
      */
     public function setStorageForImpersonator(StorageInterface $storageForImpersonator)
     {
@@ -166,7 +151,7 @@ class User extends ZfcUserUserService
      * Set the setting for storing user to the session as object (rather than ID)
      *
      * @param bool $storeAsObject
-     * @return \ZfcUser\Options\ModuleOptions
+     * @return User
      */
     public function setStoreUserAsObject($storeAsObject)
     {
